# openapi.yml
openapi: 3.0.3
info:
  title: Easy!Appointments API
  version: 1.0.0
  description: These are the OpenAPI specs that describe the REST API of Easy!Appointments.
  termsOfService: https://easyappointments.org/terms-and-conditions
  contact:
    name: Easy!Appointments
    url: https://easyappointments.org
    email: info@easyappointments.org
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
servers:
  - url: http://localhost/api/v1
    description: Ambiente local (Docker)
  - url: http://localhost/index.php/api/v1
    description: Ambiente local — URL alternativa com index.php
  - url: https://demo.easyappointments.org/index.php/api/v1/
    description: Demo
...
paths:
...
  /companies/{companyId}:
    get:
      tags:
        - companies
      summary: Get a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyRecord'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
    put:
      tags:
        - companies
      summary: Update a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyPayload'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyRecord'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codegen-request-body-name: body
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
    delete:
      tags:
        - companies
      summary: Delete a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerToken: [ ]
        - BasicAuth: [ ]
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      example:
        success: false
        message: Test error message with debug info.
    CompanyRecord:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
          description: URL-friendly unique identifier for the company.
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        logo:
          type: string
          nullable: true
        timezone:
          type: string
        language:
          type: string
          nullable: true
      example:
        id: 1
        name: EMPRESA A
        slug: empresa-a
        description: null
        address: Rua Exemplo, 123
        city: Recife
        state: PE
        zip: '50000-000'
        phone: '+558100000000'
        email: contato@empresaa.com
        notes: null
        logo: null
        timezone: America/Recife
        language: portuguese-br
    CompanyPayload:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
        slug:
          type: string
          description: URL-friendly unique identifier (must be unique across companies).
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        timezone:
          type: string
        language:
          type: string
          nullable: true
      example:
        name: EMPRESA A
        slug: empresa-a
        email: contato@empresaa.com
        timezone: America/Recife
...
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      description: >-
        You can generate the API key from the settings page to authorize with
        Bearer Token.
    BasicAuth:
      type: http
      scheme: basic
      description: >-
        You can use the credentials of an admin user to authorize with Basic
        Auth.